refactored app.js into Home.js
